Grind main/partyobj: signed modulo in bonus frame check (81.04% -> 81.53%)
System.m_frameCounter is unsigned, but the original % 30 check used a
signed modulo. Cast to int so MWCC emits the target's signed
divide-by-constant (subi 0x7777 / mulhw / srawi) instead of the unsigned
sequence. The remaining mismatch is register coloring.

+++ src/partyobj.cpp
@@ -3475,7 +3475,7 @@ void CGPartyObj::bonus(int kind, int value, CGPrgObj* source)
 	default:
 		break;
 	case 0:
-		if (kind == 2 && (System.m_frameCounter % 30) == 0) {
+		if (kind == 2 && (static_cast<int>(System.m_frameCounter) % 30) == 0) {
 			addValue = stageAdd;
 		}
 		break;
